# docs/swagger-1.1.yaml
swagger: '2.0'
info:
  description: Pho Networks REST API
  version: "1.1.0"
  title: Social Graph API
  contact:
    email: emre@phonetworks.org

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /{uuid}/edges/all:
    get:
      operationId: getAllEdges
      summary: retrieves the edges of a node
      description: |
        By passing in a node ID, you can fetch all
        the edges of the node in question; including incoming and
        outgoing.
      produces:
      - application/json
      parameters:
      - in: path
        name: uuid
        description: The node ID
        required: true
        type: string
      responses:
        200:
          description: search results matching criteria
          schema:
              type: object
              required:
              - from
              - to
              - in
              - out
              properties: 
                from:
                  type: array
                  items:
                    $ref: '#/definitions/NodeEdge'
                to:
                  type: array
                  items:
                    $ref: '#/definitions/NodeEdge'
                in:
                  type: array
                  items:
                    type: object
                    required:
                    - "5f9e769ff6fe4cee02b3b4547200db6b"
                    properties:
                      "5f9e769ff6fe4cee02b3b4547200db6b":
                        $ref: '#/definitions/NodeEdge'
                out:
                  type: array
                  items:
                    type: object
                    required:
                    - "5f9e769ff6fe4cee02b3b4547200db6b"
                    properties:
                      "5f9e769ff6fe4cee02b3b4547200db6b":
                        $ref: '#/definitions/NodeEdge'
        500:
          description: Failure.
  /{uuid}/edges/in:
    get:
      operationId: getIncomingEdges
      summary: retrieves the incoming edges of a node
      description: |
        By passing in a node ID, you can fetch 
        the incoming edges of the node in question.
      produces:
      - application/json
      parameters:
      - in: path
        name: uuid
        description: the node ID
        required: true
        type: string
      responses:
        200:
          description: The incoming edges.
          schema:
            type: array
            items:
              $ref: '#/definitions/NodeEdge'
        500:
          description: Failure.
  /{uuid}/edges/out:
    get:
      operationId: getOutgoingEdges
      summary: retrieves the outgoing edges of a node
      description: |
        By passing in a node ID, you can fetch 
        the outgoing edges of the node in question.
      produces:
      - application/json
      parameters:
      - in: path
        name: uuid
        description: the node ID
        required: true
        type: string
      responses:
        200:
          description: The outgoing edges.
          schema:
            type: array
            items:
                $ref: '#/definitions/NodeEdge'
        500:
          description: Failure.
...
definitions:
...
  NodeEdge:
    type: object
    required:
    - id
    - classes
    properties:
      id:
        type: string
        format: uuid
        example: 8b79d958f3bb5b625808f490a3ec0ed6
      classes:
        type: array
        items:
          type: string
        example: [
                "PhoNetworksAutogenerated\\UserOut\\Post",
                "Pho\\Kernel\\Foundation\\ActorOut\\Write",
                "Pho\\Framework\\ActorOut\\Write",
                "Pho\\Framework\\ActorOut\\Subscribe",
                "Pho\\Framework\\AbstractEdge",
                "Pho\\Lib\\Graph\\Edge"
          ]
...
host: virtserver.swaggerhub.com
basePath: /phonetworks/server-rest/1.0.0
schemes:
 - https